Edit UpdateChar Method

+++ Services/CharacterService/CharacterService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_rpg2.Data;
@@ -13,20 +14,27 @@ namespace dotnet_rpg2.Services.CharacterService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         } 
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
             serviceResponse.Data = await _context.Characters
+                .Where(c => c.User.Id == GetUserId())
                 .Select(c => _mapper.Map<GetCharacterDto>(c))
                 .ToListAsync();
             return serviceResponse;
@@ -38,12 +46,22 @@ namespace dotnet_rpg2.Services.CharacterService
 
             try
             {
-                var character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                response.Data = _context.Characters
-                .Select(c=> _mapper.Map<GetCharacterDto>(c))
-                .ToList();
+                var character = await _context.Characters
+                    .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                    if(character != null)
+                    {
+                         _context.Characters.Remove(character);
+                        await _context.SaveChangesAsync();
+                        response.Data = _context.Characters
+                            .Where(c => c.User.Id == GetUserId())
+                            .Select(c=> _mapper.Map<GetCharacterDto>(c))
+                            .ToList();
+                    }
+                    else
+                    {
+                        response.Success = false;
+                        response.Message = "Character not found.";
+                    }        
             } 
             catch(Exception ex)
             {
@@ -56,7 +74,9 @@ namespace dotnet_rpg2.Services.CharacterService
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var response = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _context.Characters.ToListAsync();
+            var dbCharacters = await _context.Characters
+            .Where(c => c.User.Id == GetUserId())
+            .ToListAsync();
             response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return response;
         }
@@ -64,7 +84,9 @@ namespace dotnet_rpg2.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var response = new ServiceResponse<GetCharacterDto>();
-            var dbcharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbcharacter = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+
             response.Data = _mapper.Map<GetCharacterDto>(dbcharacter);
             return response;
         }
@@ -75,8 +97,11 @@ namespace dotnet_rpg2.Services.CharacterService
 
             try{
             var character = await _context.Characters
+                .Include(c => c.User)
                 .FirstOrDefaultAsync(c => c.Id == updated.Id);
 
+            if(character.User.Id == GetUserId())
+            {
             character.Name = updated.Name;
             character.HitPoints = updated.HitPoints;
             character.Strength = updated.Strength;
@@ -87,6 +112,11 @@ namespace dotnet_rpg2.Services.CharacterService
             await _context.SaveChangesAsync();
 
             response.Data = _mapper.Map<GetCharacterDto>(character);
+            }
+            else {
+                response.Success = false;
+                response.Message = "Character not found.";
+            }
             } 
             catch(Exception ex)
             {
@@ -95,5 +125,7 @@ namespace dotnet_rpg2.Services.CharacterService
             }
             return response;
         }
+
+
     }
 }
